refactor: rename ui to common
