+++ src/app/blog/[slug]/page.tsx
@@ -2,13 +2,13 @@ import type { Post } from 'contentlayer/generated'
 import { allPosts } from 'contentlayer/generated'
 import { notFound } from 'next/navigation'
 
+import { Container } from '@/components/common'
 import ContentMeta from '@/components/content-meta'
 import Link from '@/components/link'
 import Mdx, { Image } from '@/components/mdx'
 import PageHeader from '@/components/page-header'
 import Reactions from '@/components/reactions'
 import ShareButton from '@/components/share-button'
-import { Container } from '@/components/ui'
 import cn from '@/lib/cn'
 import { getJsonLd, getMetadata } from '@/lib/metadata'
 import { formatDate, getBaseUrl, kebabCase } from '@/lib/utils'
