+++ src/app/snippets/[slug]/page.tsx
@@ -2,12 +2,12 @@ import type { Snippet } from 'contentlayer/generated'
 import { allSnippets } from 'contentlayer/generated'
 import { notFound } from 'next/navigation'
 
+import { Container } from '@/components/common'
 import ContentMeta from '@/components/content-meta'
 import Mdx from '@/components/mdx'
 import PageHeader from '@/components/page-header'
 import Reactions from '@/components/reactions'
 import ShareButton from '@/components/share-button'
-import { Container } from '@/components/ui'
 import cn from '@/lib/cn'
 import { getJsonLd, getMetadata } from '@/lib/metadata'
 import { getBaseUrl } from '@/lib/utils'
