+++ src/app/snippets/page.tsx
@@ -2,9 +2,9 @@ import type { Snippet } from 'contentlayer/generated'
 import { allSnippets } from 'contentlayer/generated'
 import type { Metadata } from 'next'
 
+import { Container, EmptyState } from '@/components/common'
 import Link from '@/components/link'
 import PageHeader from '@/components/page-header'
-import { Container, EmptyState } from '@/components/ui'
 import { ROUTES } from '@/constants/links'
 import cn from '@/lib/cn'
 import { getMetadata } from '@/lib/metadata'
