+++ src/components/common/counter.tsx
