+++ src/components/endorsements.tsx
@@ -7,9 +7,9 @@ import { useEndorsements } from '@/hooks'
 import cn from '@/lib/cn'
 import type { SkillCategory } from '@/types/skills'
 
+import { Container, Spinner } from './common'
 import EndorsementsBadge from './endorsements-badge'
 import Link from './link'
-import { Container, Spinner } from './ui'
 
 interface EndorsementsProps {
   fallbackData: Array<SkillCategory>
