+++ src/components/post-card.tsx
@@ -10,9 +10,9 @@ import cn from '@/lib/cn'
 import { max } from '@/lib/screens'
 import { formatDate } from '@/lib/utils'
 
+import { IncrementCounter, Spinner } from './common'
 import { Clock, Eye } from './icons'
 import Link from './link'
-import { IncrementCounter, Spinner } from './ui'
 
 type LayoutOption = 'list' | 'grid'
