+++ src/components/reactions.tsx
@@ -8,8 +8,8 @@ import { MAX_REACTIONS_PER_SESSION } from '@/constants/reactions'
 import { useReactions } from '@/hooks'
 import cn from '@/lib/cn'
 
+import { Counter } from './common'
 import EmojiReaction from './emoji-reaction'
-import { Counter } from './ui'
 
 interface ReactionsProps {
   slug: string
