+++ src/components/recent-posts.tsx
@@ -3,10 +3,10 @@ import { allPosts, type Post } from 'contentlayer/generated'
 import { ROUTES } from '@/constants/links'
 import cn from '@/lib/cn'
 
+import { Button, EmptyState, Section } from './common'
 import { ChevronRight } from './icons'
 import Link from './link'
 import PostCard from './post-card'
-import { Button, EmptyState, Section } from './ui'
 
 const getRecentPosts = (maxDisplay: number = 2) =>
   allPosts
